LiveSplit Auto Splitting Runtime Proof of Concept
I'm not sure if you guys are aware, but we've been working on developing
the "next generation" of auto splitting in LiveSplit recently. The idea
is to do it right this time, solving the following problems:
1. We want the runtime to be portable across different operating
   systems, such as Windows, Linux, and macOS.
2. The runtime should be sandboxed. It shouldn't be possible for an auto
   splitter to steal your passwords or delete your files.
3. The language you write your auto splitter in should be your choice.
4. The new runtime shouldn't be allowed to work with LiveSplit's
   internals (which made it hard for LiveSplit to ever change its
   internals). So the new API needs to be cleanly defined and be
   portable to other timers.

All of this already works, but due to the sandboxing approach, not every
auto splitter can be rewritten for the new runtime yet. Unfortunately
with the way ASL works (accessing internals and being a C# dialect),
it has turned out to be quite hard to port it to the new runtime. So for
now I've actually looked into JavaScript / TypeScript to be the
successor language for ASL. I've however also considered Python and Lua.
In the end we'll probably support all of them anyway. So I really like
that you guys are exploring Lua here.

I've been curious how easy it would be to plug the auto splitting
runtime into LAST, so I've written up a branch that provides a Proof of
Concept implementation. It already works perfectly fine, but it lacks
some things that the Lua implementation didn't provide either (skipping
splits, setting game time, querying the timer state, ...).

If you want you can plug in this runtime, which will automatically make
LAST compatible with all future LiveSplit auto splitters. I also still
encourage you to work on the Lua auto splitters, as I believe that's
indeed a great language that we probably want to support in the runtime
as well. In fact it would be great if we worked out the design of the
Lua auto splitters such that LAST's lua auto splitters can be compatible
with LiveSplit('s auto splitting runtime) as is.

Keep up the great work :)